messaging_app/requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10
sqlparse==0.5.4
